Adjusted setting for header
