+++ App.js
@@ -3,12 +3,22 @@ import { StyleSheet, Text, View } from 'react-native';
 import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import AppNavigation from './Navigation';
+import { createContext, useState } from 'react';
+
+// Create the context
+export const AppContext = createContext();
+
 
 export default function App() {
+  const [someState, setSomeState] = useState([]);
+
   return (
-    <NavigationContainer>
-      <AppNavigation />
-    </NavigationContainer>
+    // Wrap the app with the Provider to share the state
+    <AppContext.Provider value={{ someState, setSomeState }}>
+      <NavigationContainer>
+        <AppNavigation />
+      </NavigationContainer>
+    </AppContext.Provider>
   );
 }
