+++ Screens/Activities.js
@@ -1,13 +1,22 @@
 import { StyleSheet, Text, View } from 'react-native'
 import React from 'react'
-import Header from '../Header'
+import Header from '../Components/Header'
 
 export default function Activities() {
   return (
-    <View>
-      <Text>Activities</Text>
+    <View style={styles.container}>
+      <Header
+        title='Activities'
+        rightButtonText='Add'
+        onRightButtonPress={() => navigation.navigate('AddActivity')}
+      />
     </View>
   )
 }
 
-const styles = StyleSheet.create({})
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        backgroundColor: '#D8BFD8',
+    },
+})
